Add TaskContentEventObserver & TaskContentEventSubject. Add TaskCompletionRateEventObserver & TaskCompletionRateEventSubject. Add icons folder, re-config log4j2. Fix bugs

+++ src/main/java/org/swing/app/view/components/ui/button/Button.java
@@ -11,7 +11,7 @@ import java.util.Iterator;
 
 public abstract class Button extends SimpleComponent {
 
-    public void setIcon(String iconLocation, short iconWidth, short iconHeight) {
+    public void setIcon(String iconLocation, int iconWidth, int iconHeight) {
         final ImageIcon imageIcon = ViewUtil.getImageIcon(iconLocation, iconWidth, iconHeight);
         ((AbstractButton) this.sourceComponent).setIcon(imageIcon);
     }
